+++ 2_zenodo/job.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 #SBATCH --job-name=zenodo
 #SBATCH --nodes=1
-#SBATCH --ntasks-per-node=1
+#SBATCH --ntasks-per-node=4
 #SBATCH --cpus-per-task=1
 #SBATCH --time=1:00:00
 #SBATCH --mem=5G