+++ CHANGELOG.md
@@ -17,6 +17,7 @@ A major refactoring of the ACID dataset.
   The encoding/decoding is stored in a file as part of the data generation.
 - Data is stored in ordinary ZIP files, mostly compatible with the NPZ format.
   They also include a `meta.json` file with some metadata.
+- Data is now uploaded to Zenodo as one ZIP file per kernel instead of a single ZIP file containing the full dataset.
 
 ### Removed